+++ src/ReinforcementLearningEnvironments/src/environments/classic_control/pendulum.jl
@@ -9,7 +9,7 @@ struct PendulumEnvParams{T}
     m::T
     l::T
     dt::T
-    max_steps::Int64
+    max_steps::Int
 end
 
 mutable struct PendulumEnv{T, R<:AbstractRNG} <: AbstractEnv
@@ -18,7 +18,7 @@ mutable struct PendulumEnv{T, R<:AbstractRNG} <: AbstractEnv
     observation_space::MultiContinuousSpace{(3,), 1}
     state::Array{T, 1}
     done::Bool
-    t::Int64
+    t::Int
     rng::R
 end
